Update patch set 1
Patch Set 1: Code-Review-1

I think I might not understand the code around line 805. Specifically, the "non-pointer" comment makes me distrustful. Consider that this is an else from

    else if et.Kind() == reflect.Ptr && !existing.IsNil()

What if et.Kind() is reflect.Ptr, but existing.IsNil()? We end up in this case, and we have a comment saying "it's a non-pointer, we'll copy it and decode into it." And I distrust this.

So even if the code's right, I would want that comment to be less scary. My intuition is that you don't want to combine those tests; if it's a pointer, we should either use it (if it's not nil) or make a new thing, and if it's not a pointer, we should copy-and-reuse. As-is, I think that if it's a nil pointer, subv ends up being the result of calling new(*ptr) and dereferencing it, yielding another nil pointer of the same type which was separately allocated?

(But don't rely too heavily on my reading of the code, I am not good at reflect.)

Patch-set: 1
Reviewer: Gerrit User 29554 <29554@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1