Update patch set 2
Patch Set 2:

The code looks reasonable to me. That said, since there was a bug in the previous patch set, I'd suggest also having a test that would have caught that bug; that way, we confirm that it's fixed (it should be, I think), and also document that quality of the system so the next person working on this this code doesn't make the same mistake?

Patch-set: 2
Reviewer: Gerrit User 29554 <29554@62eb7196-b449-3ce5-99f1-c037f21e1705>