cmd/go: recognise illumos suffix when constructing build list
Similar to #30888, but for illumos/solaris instead of android/linux.
With this change, imports.MatchFile understands that when the
"illumos" tag is set, the "solaris" tag is implied.

In addition, cmd/go/internal/imports/tags.Tags() constructs a map of
active build tags.  By default this includes a tag for GOOS, GOARCH,
etc.  When building for the "illumos" target we need to include the
implied "solaris" tag as well.

Fixed #32370

Change-Id: I5be315c5584c694038c2feb27113809eac1e0579