runtime: change the page size to 64 KiB
DO NOT SUBMIT

This change makes the size of a Go runtime page 64 KiB instead of 8 KiB
and makes it so that the largest "small" object is 256 KiB instead of 32
KiB.

Because the largest "small" object is now greater than the oblet size,
some assumptions in scanobject no longer hold. Namely, the bases of objects
larger than the oblet size don't necessarily correspond to the beginning
of a span. As a result, we modify scanobject to compute the base of the
object accordingly in cases where the object is greater than the oblet
size, but smaller than the largest "small" object.

Change-Id: Ib48278da2e3b9268268a1cf491e46577c1b47ce4