Update patch set 1
Patch Set 1:

Thanks for sending a change.

The upstream issue is still in NeedsInvestigation stage (not yet in NeedsFix one), so it'd be helpful to discuss the plan for how to resolve the issue there first. Once there's an agreement there, we can review the code that implements the fix.

Patch-set: 1
CC: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>