vcs: Add support for opendev.org
Support for OpenStack was added in Ia64bdb1dafea33b1c3770803230d30ec1059df22,
and there are still projects working on implementing portions of their
functionality in go.

Since that was added though, the infrastructure services for the OpenStack
community have rebranded to be OpenDev, with an expanded scope to support other
open source projects in addition to just the OpenStack project. This is
expected to bring in a more diverse mix of projects and technology choices.

As part of that rebranding, the git services were updated from being at
git.openstack.org to being hosted at opendev.org. The existing support for
git.openstack.org should be kept for some time to make sure there is backwards
compatibility. The opendev team have committed to providing redirects from the
old URLs to the new, and there may be various places where things are hard
coded to git.openstack.org. That support can likely be removed in the future,
but for now opendev.org should be added in addition to the legacy URL support.

Fixes #32627

Change-Id: I5e528960c62d2c487e9de4ac136eba9765f064cb