Update patch set 3
Patch Set 3:

There has been some question as to whether this is actually needed or not with the code automatically trying different methods to connect to the repo. See experiment here:

http://paste.openstack.org/show/752934/

Just let me know if this should be abandoned, or if it should be changed to just remove the existing git.openstack.org.

Patch-set: 3
Reviewer: Gerrit User 32572 <32572@62eb7196-b449-3ce5-99f1-c037f21e1705>