Update patch set 1
Patch Set 1:

(2 comments)

Thank you for the change Raghavendra and for the ping Agniva!

Raghavendra, nice! LGTM but with some suggestions to revert the "maximum" in
IdleTimeout as well as to update the commit message. I'll +1 it and when the
changes are made, I'll give the +2. Thanks!

Patch-set: 1
Reviewer: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>