src: apply tryhard -err="" -r $GOROOT/src
This CL gives an impression of the std library written using `try`.
Some transformations may be incorrect due to the limitations of the
tool (see https://github.com/griesemer/tryhard). Some transformations
may be missing.

Even with `try` we won't be able to rewrite a significant chunk of
the std library since parts of it need to compile with Go 1.4 for
bootstrapping. Just something to keep in mind when looking at this
code.

DO NOT REVIEW.
DO NOT SUBMIT.

Updates #32437.

Change-Id: Ib9f8de8ad03d9ba2176d66363280f4f0c57499ef