runtime: fix SetMaxHeap semantics
This change re-aligns SetMaxHeap with its documented semantics, that is,
it will send a notification for each time the result of ReadGCPolicy()
changes.

Currently, it is correct for changes to GOGC and effective GOGC, but not
when maxHeap is changed. This is because SetMaxHeap sets
gcPressure.maxHeap, and ReadGCPolicy (and consequently
gcSetTriggerRatio) reads from gcPressure.maxHeap.

This change fixes the issues with semantics by also maintaining a global
maxHeap variable (much like the global gcpercent variable) which
represents the most up-to-date version of the max heap.

In short, maxHeap is to gcPressure.maxHeap as gcpercent is to
gcPressure.gogc. The former are the "most up-to-date" while gcPressure
always contains the last policy gcSetTriggerRatio saw.

This change also updates the SetMaxHeap test to check for a notification
after the first call to SetMaxHeap, not the second. This helps ensure
the behavior of the notification channel that's documented is actually
maintained.

Furthermore, this change makes some parts of SetMaxHeap that grab the
heap lock, grab it on the system stack to avoid a deadlock with stack
growth.

Change-Id: I93409f40d255b95b5fc1c0ec26d5e1b3d5a73849