Update patch set 1
Patch Set 1:

Presumably this improves performance - if so, add benchmark numbers to the commit message.

If performance isn't affected, I'm not sure if this is worth doing. The old code is just as fine, and even safer, as we can more easily prove that the pool always provides states which have been reset.

Patch-set: 1