[RFC]sync/atomic: specify the memory order guarantee provided by Load/Store
Current Go documentation does not specify any memory order guarantees
made by the atomic operations, but implementations of sync/atomic are
concerned with the ordering. So it is better to state the behavior of
the atomic operations wrt the memory model to avoid wrong assumptions
being made by Go developers.

The compiler emits different instructions for atomic.Load and atomic.Store
on different architectures, as summarized below.
architecture      atomic.Load      atomic.Store
AMD64                load             XCHG
S390x                load           store+SYNC
MIPS64         SYNC+load+SYNC      SYNC+store+SYNC
PPC64          SYNC+load+ISYNC      SYNC+load
ARM64                LDAR             STLR

The atomic.Store is generated as XCHG on AMD64 and as "store+SYNC" on S390x
to prevent store-load reordering, and the atomic.Load is implemented as
normal load because other sequences of memory operations (load-load,
store-store and load-store) are already guaranteed by hardware not to be
reordered.

The behavior of atomic operations is impacted not only by the instructions
generated for each CPU, but also by the behavior of the compiler. The
compiler behavior of atomic.Load is not sequentially consistent, because
if there is no dependency between normal load IR and later atomic load IR,
the order is not guaranteed to be preserved during compilation.

The behavior of the compiler and CPU combined, atomic.Load and atomic.Store
are not sequentially consistent. And the CPU behavior on arm64 is "Release"
semantics for atomic.Store and "Acquire" semantics for atomic.Load. Considering
the current implementation, it might be reasonable that we just refer to the
most relaxed memory ordering behavior, so this CL specifies that atomic.Load
matches acquire semantics and atomic.Store matches release semantics (like
c++ sequentially consistent acquire and release memory order).

Update #5045

Change-Id: I10ffa4ac016070346ff03ae6193beeb1cc94626d