cmd/doc: add option to suppress showing package docs
Currently there is no way for go doc to output a clean single line separated symbole definitions of types, functions, vars and consts without documentation noise.

For example `go doc | fzf` the intro of a package is scrambled between the output for a fuzzy finder.

Added a `-short` flag to be able to do `go doc -short | fzf`

Fixes #32597

Change-Id: I77a73838adc512c8d1490f5a82075de6b0462a31