encoding/json: correct caller's name in encoding errors
1. mapEncoder.encode does not use MarshalerError any more.
MarshalerError's Error method refers to MarshalJSON, but that method was never
called in mapEncoder.encode. Now is using simple fmt.Errorf without
confusing error message.

2. MarshalerError's message always refers to MarshalJSON, while it can
be called from MarshalText. And in the last case the error message looks
not very clear. Was added a new struct MarshalerErrorExt with extra
field where can be stored the name of the original function. All places
where MarshalerError was used where changed to use new struct.
We could not change MarshalerError because it's a public type and it's
change may break user's code.

Fixes #29753

Change-Id: I186c2fac8470ae2f9e300501de3730face642230