runtime: block on background scavenger only if necessary
This change makes it so that an allocation will only ever block on the
background scavenger if the allocation tries to take the same memory
that's currently being scavenged. This only applies to the background
scavenger and not to debug.FreeOSMemory or the allocation-time
scavenging.

To facilitate this improvement, the background scavenger does everything
it possibly can with the to-be-scavenged span except for making the
lengthy syscall. It then leaves a note describing which memory range its
scavenging, releases the heap lock, and continues scavenging.

Allocations which attempt to allocate from scavenged memory then check
the note left behind, and if the memory region they're acquiring
overlaps, then they wait. The background scavenger will signal the waiter
when it's done.

Note that if an allocation ends up getting blocked on the background
scavenger, it blocks all other allocations because it continues to hold
the heap lock. This change merely reduces the number of cases in which
the background scavenger forces all allocations to wait, but does not
eliminate the chance.

Change-Id: I187c0a87f1f2e72be1e17b56e81e2e78532ef26f