runtime: unscavenge span without heap lock
If we go to sleep because the background scavenger is scavenging a span
that we currently want to allocate out of, then anything that needs to
acquire the heap lock will queue up behind us.

This change adds a new unscavenge method to mspan which calls sysUsed on
the span once its safe to do so (that is, once the background scavenger
is done, if it happens to be touching this span). This change also
changes the contract of allocSpanLocked to require the caller to
unscavenge (though not to perform memstats updates).

As a result, unscavenging, and thus waiting for the background
scavenger, can happen with the heap unlocked, side-stepping this
problem. A consequence of this then is that multiple allocators could
end up waiting on the background scavenger, such as if the scavenger is
scavenging an 8-page span that was allocated in 1-page increments by 8
other Ps while the background scavenger was working. To allow one waiter
per pending span, we add a note to each span and maintain a waiters list
(represented as a linked list of spans) instead of a single global note.
The background scavenger now walks over the list and wakes up each
waiter.

A nice bonus is that unscavenging now happens with the heap unlocked, so
systems which actually do something in sysUsed (like Windows) no longer
need to pay a system-wide penalty for this.

Change-Id: I7850cb85857caf5ddf9ac3a9d566440e5e5a55a5