wrapper: built-in wrapper interface
Unwrap, Is, As, %w can be abstracted for other type's too, not only for errors. This attempt tries to introduces a built-in wrapper interface.

I think this decision needs to be made before the new error packages goes life because don't think this can be made backward compatible once the new error package goes life with it's locked in type error in wrap.go

Fixes: #32913
Change-Id: I29ae650498cd496c8c293b58f0f70c62690c1b2b