cmd/go: automatically call testing.Init under 'go test'
In Go 1.12 and earlier, the testing package initialized its flags at
init time. This meant that non-test code had to be careful never to
import testing lest it pull in those irrelevant flags; see #21051.

To fix this, in CL 173722 we moved the flag registration to testing.Init
which was called automatically in testing.MainStart.

Then we discovered that people were calling flag.Parse in init functions
in their tests, so in CL 176098 we injected code to call testing.Init
into the test package.

Unfortunately, as outlined in #31859, it turns out that there are even
non-test packages in the wild that call flag.Parse in init functions and
those packages may be dependencies of tests. (This is clearly
ill-advised: if such a package happens to be initialized before testing,
which could happen if that package does not itself depend on testing, it
will break the testing flag registration even in Go 1.12 and earlier.
Whether or not it breaks depends on unspecified, compiler-specific
initialization ordering.)

However, we can prevent even this breakage. This CL removes the
"testinginit" workaround added in CL 176098 and adds a file, init.go,
to the the testing package which calls Init at initialization time.
This file has an 'ignore' build constraint, but cmd/go un-ignores
init.go as a special case under 'go test'.

Fixes #31859

Change-Id: I1410aa04cb5bb320bc90570a6f64ab9668e00ed2