Update patch set 1
Patch Set 1: Code-Review-1

Thank you for working on this Cuong!

While the end goal of having these examples run on the playground
is great, losing precision and focus on the examples isn't. If
an example for ReadFile also requires creating a temporary file
and then an os.Remove means that folks can't copy and paste the
example anymore :( For that reason, I am going to vote a -1

Patch-set: 1
Reviewer: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1