Update patch set 2
Patch Set 2:

I'm personally a bit uneasy about this hidden magic. I understand the old code already changed the 'go test' build, but at least it changed the temporary test package (which already contains generated code) and not a package in the standard library which is otherwise static.

For example, if one runs make.bash and then runs a test, will the testing package be rebuilt? I assume it will, as make.bash has only filled the build cache with the non-test version of the testing package.

Patch-set: 2