Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I'm personally a bit uneasy about this hidden magic. I understand the old code already changed the 'go test' build, but at least it changed the temporary test package (which already contains generated code) and not a package in the standard library which is otherwise static.

The old version only helps if the init which calls flag.Parse is in the test package itself, not if it's in a dependency. So I'm not sure there's much value in having such a workaround.

I think we should either get this working (i.e., make the breakage ~invisible), or else go back to the state in CL 173722 and just tell people that calling flag.Parse from init is not a supported thing.

Patch-set: 2
Reviewer: Gerrit User 5615 <5615@62eb7196-b449-3ce5-99f1-c037f21e1705>