Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > How about we keep everything as is, but just do an os.MkdirAll('testdata', 0644) before reading/writing the file ?
> > > 
> > > That should make it work in the playground and make a very minimal change.
> > 
> > 
> > It would work for WriteFile, but not for ReadFile, you still have to creat the file.
> 
> Indeed. I don't have a better idea for ReadFile. Even os.Open() does not have an example.

The WriteFile example should not be writing to the 'testdata' directory anyway; see #28387.

I also don't understand why `golang.org/pkg/io/ioutil' is presenting the examples as a `func main()` in the first place. There is no general requirement that `Example` functions be runnable as standalone `main` functions; if we want to convert Example functions to run in the Playground we should arguably convert them as the entire _test.go file, perhaps with the other tests and examples in the file redacted out and a fresh `goimports` run to clear out any additional dependencies that are no longer needed.

Patch-set: 1
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>