runtime: avoid re-acquiring heap lock in scavengeLocked
This change re-arranges scavengedLocked for the background scavenger to
instead only enqueue memory regions in a global queue, instead of making
the syscall inline and having to reacquire the heap lock. Any allocator
must now look over the whole queue instead of just a single pair of
notes, but it will always look at at most the length of the queue, which
is very small.

Change-Id: Ib0b7c00cccfbe32caf837bfd028ec56f5c3aad98