Update patch set 1
Patch Set 1:

> > Thanks, but let's not write this kind of specification in a CL.  Let's first decide, on the issue or the mailing list, what docs we want.
> 
> The issue (https://github.com/golang/go/issues/5045) on how to define atomic memory model was posted six years ago and there is no clear result until now. I also submitted a comment under this issue, but I didn't get any answers, so we push this CL for discussion and want to get some advices. I will copy the commit message to the mailing list for further discussion. Thank you.

I appreciate that this is a frustrating process.  For various reasons few people on the core Go team are interested in tackling this, as it is complicated, confusing, and doesn't really matter to 99% of Go programmers.  As the memory spec says: "If you must read the rest of this document to understand the behavior of your program, you are being too clever."

Despite all that, the way to solve this problem is to discuss it on the issue.  Sending a patch is just distracting and does not move us forward.

Thanks.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>