Update patch set 19
Patch Set 19:

> Added one more in `const predeclReserved = 33` 

This is fine to leave at 32. It's just that the first 32 types are reserved, not that they're all used for predeclared types.

You need to add an additional type to the "predeclared" list in go/internal/gcimporter/bimport.go.

Sorry for confusingly specifying iimport.go in my last message. I mentioned it only because it's what was listed by "git grep" for the error string, and I forgot gcimporter still had the bimport.go/iimport.go split structure too.

Patch-set: 19